+++ ContentPatcher/Framework/Patches/Patch.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using ContentPatcher.Framework.Conditions;
+using ContentPatcher.Framework.Lexing.LexTokens;
 using ContentPatcher.Framework.Tokens;
 using StardewModdingAPI;
 
@@ -19,6 +20,9 @@ namespace ContentPatcher.Framework.Patches
         /// <summary>The underlying contextual values.</summary>
         protected readonly List<IContextual> ContextualValues = new List<IContextual>();
 
+        /// <summary>The context which provides tokens for this patch, including patch-specific tokens like <see cref="ConditionType.Target"/>.</summary>
+        protected SinglePatchContext PrivateContext { get; }
+
 
         /*********
         ** Accessors
@@ -28,9 +32,6 @@ namespace ContentPatcher.Framework.Patches
 
         /// <summary>Whether the instance is valid for the current context.</summary>
         public bool IsReady { get; protected set; }
-
-        /// <summary>The last context used to update this patch.</summary>
-        protected IContext LastContext { get; private set; }
 
         /// <summary>A unique name for this patch shown in log messages.</summary>
         public string LogName { get; }
@@ -65,34 +66,42 @@ namespace ContentPatcher.Framework.Patches
         /// <returns>Returns whether the patch data changed.</returns>
         public virtual bool UpdateContext(IContext context)
         {
-            this.LastContext = context;
+            bool isReady = true;
             bool changed = false;
+
+            // update target asset (needed by patch context)
+            this.RawTargetAsset.UpdateContext(context);
+            this.TargetAsset = this.NormaliseAssetName(this.RawTargetAsset.Value);
+
+            // update patch context
+            this.PrivateContext.Update(context, this.RawTargetAsset);
 
             // update contextual values
             foreach (IContextual contextual in this.ContextualValues)
             {
+                if (contextual == this.RawTargetAsset)
+                    continue; // updated above
+
                 bool wasReady = contextual.IsReady;
-                if (contextual.UpdateContext(context) || contextual.IsReady != wasReady)
+                if (contextual.UpdateContext(this.PrivateContext) || contextual.IsReady != wasReady)
                     changed = true;
             }
 
             // update source asset
             if (this.FromLocalAsset != null)
             {
-                bool sourceChanged = this.FromLocalAsset.UpdateContext(context);
-                this.IsReady = this.IsReady && this.FromLocalAsset.IsReady && this.ContentPack.HasFile(this.FromLocalAsset.Value);
+                bool sourceChanged = this.FromLocalAsset.UpdateContext(this.PrivateContext);
+                isReady = this.FromLocalAsset.IsReady && this.ContentPack.HasFile(this.FromLocalAsset.Value);
                 changed = changed || sourceChanged;
             }
-
-            // update target asset
-            this.TargetAsset = this.NormaliseAssetName(this.RawTargetAsset.Value);
 
             // update ready flag
             {
                 bool wasReady = this.IsReady;
                 this.IsReady =
-                    (!this.Conditions.Any() || this.Conditions.All(p => p.IsMatch(context)))
-                    && this.GetTokensUsed().All(name => context.Contains(name, enforceContext: true));
+                    isReady
+                    && (!this.Conditions.Any() || this.Conditions.All(p => p.IsMatch(this.PrivateContext)))
+                    && this.GetTokensUsed().All(name => this.PrivateContext.Contains(name, enforceContext: true));
                 changed = changed || this.IsReady != wasReady;
             }
 
@@ -120,7 +129,26 @@ namespace ContentPatcher.Framework.Patches
         /// <summary>Get the token names used by this patch in its fields.</summary>
         public virtual IEnumerable<string> GetTokensUsed()
         {
-            return this.RawTargetAsset.GetTokensUsed();
+            // from local asset
+            if (this.FromLocalAsset != null)
+            {
+                foreach (LexTokenToken lexToken in this.FromLocalAsset.GetTokenPlaceholders(recursive: true))
+                    yield return lexToken.Name;
+            }
+
+            // raw target asset
+            foreach (LexTokenToken lexToken in this.RawTargetAsset.GetTokenPlaceholders(recursive: true))
+                yield return lexToken.Name;
+
+            // conditions
+            foreach (string name in this.Conditions.SelectMany(p => p.GetTokensUsed()))
+                yield return name;
+        }
+
+        /// <summary>Get the context which provides tokens for this patch, including patch-specific tokens like <see cref="ConditionType.Target"/>.</summary>
+        public IContext GetPatchContext()
+        {
+            return this.PrivateContext;
         }
 
 
@@ -143,6 +171,7 @@ namespace ContentPatcher.Framework.Patches
             this.RawTargetAsset = assetName;
             this.Conditions = conditions.ToArray();
             this.NormaliseAssetName = normaliseAssetName;
+            this.PrivateContext = new SinglePatchContext(scope: this.ContentPack.Manifest.UniqueID);
 
             // track contextuals
             this.ContextualValues.AddRange(this.Conditions);