+++ ContentPatcher/Framework/Tokens/ImmutableToken.cs
@@ -12,9 +12,10 @@ namespace ContentPatcher.Framework.Tokens
         /// <summary>Construct an instance.</summary>
         /// <param name="name">The token name.</param>
         /// <param name="values">Get the current token values.</param>
+        /// <param name="scope">The mod namespace in which the token is accessible, or <c>null</c> for any namespace.</param>
         /// <param name="allowedValues">The allowed values (or <c>null</c> if any value is allowed).</param>
         /// <param name="canHaveMultipleValues">Whether the root may contain multiple values (or <c>null</c> to set it based on the given values).</param>
-        public ImmutableToken(string name, InvariantHashSet values, InvariantHashSet allowedValues = null, bool? canHaveMultipleValues = null)
-            : base(new ImmutableValueProvider(name, values, allowedValues, canHaveMultipleValues)) { }
+        public ImmutableToken(string name, InvariantHashSet values, string scope = null, InvariantHashSet allowedValues = null, bool? canHaveMultipleValues = null)
+            : base(new ImmutableValueProvider(name, values, allowedValues, canHaveMultipleValues), scope) { }
     }
 }